Classes finalizadas, precisa de testes

+++ Classes/Corpo.cs
@@ -3,21 +3,21 @@ namespace TrabalhoN1.Classes
     public class Corpo
     {
         public string Nome { get; set; }
-        public float Massa { get; set; }
-        public float Raio { get; set; }
-        public int PosX { get; set; }
-        public int PosY { get; set; }
-        public int VelX { get; set; }
-        public int VelY { get; set; }
+        public double Massa { get; set; }
+        public double Raio { get; set; }
+        public double PosX { get; set; }
+        public double PosY { get; set; }
+        public double VelX { get; set; }
+        public double VelY { get; set; }
 
 
         public Corpo()
         {
             Nome = "";
         }
-        public float Aceleracao(float forca)
+        public double Aceleracao(double forca)
         {
-            float a = 0;
+            double a = 0;
             a = forca * Massa;
             return a;
         }
